pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gdsfill"
version = "0.1.0"
description = "A Python project using pyyaml, packaging, and gdstk."
authors = [
    { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"

dependencies = [
    "pyyaml",
    "packaging",
    "gdstk",
]

[project.urls]
homepage = "https://example.com"
repository = "https://github.com/yourusername/my-python-project"

[project.scripts]
gdsfill = "gdsfill.gdsfill:main"
